+++ app/graphql/mutations/registro_delete.rb
@@ -13,6 +13,10 @@ module Mutations
     def resolve(id:)
       ActiveRecord::Base.transaction do
         registro = ::Registro.find(id)
+        unless can? :destroy, registro
+          raise GraphQL::ExecutionError.new "error: unauthorized access: delete 'registro'",
+                                            extensions: { code: :unauthorized }
+        end
         unless registro.destroy
           raise GraphQL::ExecutionError.new "Error deleting registro", extensions: registro.errors.to_hash
         end
